openrisc: use shadow registers to save regs on exception
Previously, the area between 0x0-0x100 have been used as a
"scratch" memory area to temporarily store regs during
exception entry.
In a multi-core environment, this will not work.

This change is to use shadow registers for nested context.

Currently only the "critical" temp load/stores are covered,
the EMERGENCY_PRINT ones are left as is (when they are used,
it's game over anyway), they need to be handled as well in the
future.

Signed-off-by: Stefan Kristiansson <stefan.kristiansson@saunalahti.fi>
Signed-off-by: Stafford Horne <shorne@gmail.com>