irqchip: add initial support for ompic
Only IPI support is implemented so far and the whole driver
should be seen as tentative at this point.

irqchip/ompic: increase number of retries

irqchip/ompic: use irqsave spinlock

ompic_raise_softirq might be called from interrupt handlers,
so if interrupts are not disabled while waiting for the lock,
the function might be re-entered, resulting in nested
spinlocks that will cause deadlocks.

Signed-off-by: Stefan Kristiansson <stefan.kristiansson@saunalahti.fi>
Signed-off-by: Stafford Horne <shorne@gmail.com>