openrisc: add l.lwa/l.swa emulation
This adds an emulation layer for implementations
that lack the l.lwa and l.swa instructions.
It handles these instructions both in kernel space and
user space.

There's a couple todos left. The biggest one is to properly
implement adjust_pc(), but to really handle the delay slots
is rather painful. An alternative would be to just detect the
situation and bail (it's hard to put l.lwa/l.swa in delay slots,
so it's an unlikely situation).

Another todo is to make all this conditional on !CONFIG_SMP,
since the emulation only makes sense in UP configurations.

Signed-off-by: Stefan Kristiansson <stefan.kristiansson@saunalahti.fi>