openrisc: scripts: Add a util to annotate bare stack traces
The openrisc stacktrace has some code to lookup symbols but
its not working so well, or sometimes its not enabled.
This script will look through a provided System.map file
and lookup any symbols on the stack. This helps to identify
return methods and stack pointers.

There is probably something better for this but I dont know what.

Example output:
[239707.650000] sp + 160: 0xc02efdf0    init_thread_union+7664
(stackdepth 528)
[239707.650000] sp + 164: 0xc02f6ca0    clockevent_openrisc_timer
[239707.650000] sp + 168: 0xc0303096    __warned.28644
[239707.650000] sp + 172: 0x0000da03
[239707.650000] sp + 176: 0x495855e4
[239707.650000] sp + 180: 0xc02efe18    init_thread_union+7704
(stackdepth 488)
[239707.650000] sp + 184: 0x00000001
[239707.650000] sp + 188: 0xc02f6ca0    clockevent_openrisc_timer
[239707.650000] sp + 192: 0x0000da03
[239707.650000] sp + 196: 0x57a22080
[239707.650000] sp + 200: 0xc02fb27c    tick_cpu_sched
[239707.650000] sp + 204: 0xc0062738    tick_program_event()+116