selftests/futex: Fix makefile dependencies
I am working on doing selftests for openrisc and found issues with the
futex test is not building after changes to the tests source.

This issue is that the TARGETS variable used in the futex Makefile is
also used by the selftests Makefile, so when building from toplevel the
TARGETS variable is clobbered and nothing gets built. Changed the
variable name to get around that.

Also, the futex makefile did not contain dependencies for all headers,
so if we make changes to logging.h rebuild will not happen. Add headers
to fix it up.

Signed-off-by: Stafford Horne <shorne@gmail.com>