smp: Sleep on secondary wait
Currently we do a spin on secondary cpus when waiting.  This causes
issues with power consumption as well as qemu cycle burning (it starves
cpu 0 from actually being able to boot.)

openrisc: head: Move the secondary interupt handler to init section

This was it can get freed as we only need it during boot.

Signed-off-by: Stafford Horne <shorne@gmail.com>