smp: sleep instead of spin on secondary wait
Currently we do a spin on secondary cpus when waiting to boot.  This causes
issues with power consumption as well as qemu cycle burning (it starves
cpu 0 from actually being able to boot.)

The secondary interrupt handler is stored in the init section so as to
allow it to be reclaimed after boot.  However, if we ever want to
support hotplug this may need to change.

Signed-off-by: Stafford Horne <shorne@gmail.com>