irq: ompic: Switch to pass ops in atomic memory
Currently the ompic has just one data slot per CPU, this means if there
are operations pending we need to wait for those to be serviced.  I have
noticed timeouts and locking issues due to this.

This change puts pending ops into a memory bit mask which can allow
multipe pending operations to be set and serviced at a time. This is
mostly borrowed from the alpha IPI implementation.

Signed-off-by: Stafford Horne <shorne@gmail.com>