dts: openrisc: fix console params for litex
The console parameter 'ttyLX' is used so that during switch from
earlycon to the real console the litex console driver can be found and
initialized.  Before this was not set and the real console
initialization never took place.

I also cleaned up other parameters, but feel free to let me know if its
wrong.

Signed-off-by: Stafford Horne <shorne@gmail.com>