tty: litex: use byte io to work on BE and LE machines
While running on openrisc qemu I found that the 32 bit reads and writes
were not working due to endianess.  I think using all byte IO in this
driver is safe to just switch to that.

Also enable interrupts so input works.

TODO, fix a bunch of TODOs also change uart_out32 to not indicate 32 bit
since it is not longer.

Signed-off-by: Stafford Horne <shorne@gmail.com>