uart: litex: Get working on real hardware
Before this patch the earlycon was working, but there was issues
switching over to the real console.  The main issue was using LE vs BE
as explained in the previous patch.  Maybe we shouldn't have to
explicitly do that, but this fixes it.

I also tried to fix some of the TODO's and remove some of the things
that were copied from the ulite driver but dont exist on the litex uart
device.

Signed-off-by: Stafford Horne <shorne@gmail.com>