openrisc: Pretty print show_registers memory dumps
Currently show registers, print memory dumps character by character and
there is no address information, so its a bit difficult to use.  For
example before a stack dump looks as follows.

    [   13.650000] Stack:
    [   13.650000] Call trace
    [   13.690000] [<(ptrval)>] ? put_timespec64+0x44/0x60
    [   13.690000] [<(ptrval)>] ? _data_page_fault_handler+0x104/0x10c
    [   13.700000]
    [   13.700000] Code:
    [   13.700000] 13
    [   13.700000] ff
    [   13.700000] ff
    [   13.700000] f9
    [   13.710000] 84
    [   13.710000] 82
    [   13.710000] ff
    [   13.710000] bc
    [   13.710000] 07
    [   13.710000] fd
    [   13.720000] 4e
    [   13.720000] 67
    [   13.720000] 84
    [   13.720000] 62
    [   13.720000] ff
    ...

This change updates this to be more compact by printing a word at time.

    [    9.340000] Stack:
    [    9.350000] Call trace:
    [    9.350000] [<(ptrval)>] handle_mm_fault+0x324/0x7b4
    [    9.360000] [<(ptrval)>] do_page_fault+0x23c/0x47c
    [    9.360000] [<(ptrval)>] ? _data_page_fault_handler+0x104/0x10c
    [    9.370000]  c0b07e48:       c008fb3c
    [    9.370000]  c0b07e4c:       c0b07e60
    [    9.370000]  c0b07e50:       c0092448
    [    9.370000]  c0b07e54:       00000000
    [    9.380000]  c0b07e58:       c0b07eb4
    [    9.380000]  c0b07e5c:       c00b4c18
    [    9.380000] (c0b07e60:)      c0ac41dc
    [    9.390000]  c0b07e64:       c0b76114
    [    9.390000]  c0b07e68:       00000cc0
    [    9.390000]  c0b07e6c:       000181ce
    [    9.390000]  c0b07e70:       3039c000
    [    9.400000]  c0b07e74:       00000255

Now we are also printing a bit of the stack as well as the code.  The
stack is output to help with debugging.  There may be concern about
exposing sensitive information on the stack, but we are already dumping
all register content which would have similar sensitive information.  So
I am going ahead as this proves useful in investigation.

Signed-off-by: Stafford Horne <shorne@gmail.com>