csky: PCI: Define ARCH_GENERIC_PCI_MMAP_RESOURCE
This looks to be something that was left out on accident on the csky
port.  Add the definition of ARCH_GENERIC_PCI_MMAP_RESOURCE to keep it
inline with other ports.  Adding ARCH_GENERIC_PCI_MMAP_RESOURCE enables
the definition of mordern pci_mmap_resource_range functions which is
generally what new ports should be using.

Acked-by: Guo Ren <guoren@kernel.org>
Signed-off-by: Stafford Horne <shorne@gmail.com>