openrisc: Add hotplug support to allow kexec with SMP
1. We update ompic to allow registering the ipi irq with the smp
intrastructure.  This allows us to shutdown ipi's to cpu's
during shutdown.

2. We move head secondary boot routines out of the .init section
to allow them to be used for hotplug.

We add __cpu_disable and idle routines to all cpus to be stopped.

The wakeup IPI is used to catch when CPUs are coming online and
redirect them to the secondary boot routine.

Signed-off-by: Stafford Horne <shorne@gmail.com>