openrisc: Add the De0 Nano SMP system
This FPGA board SoC is based on the Simple SMP reference board.  We
split out the simple_smp device tree to a sharable simple_smp.dsti
file to be able to re-use the base system definitions.

A default config is added that brings together the device tree and
driver setup.

Signed-off-by: Stafford Horne <shorne@gmail.com>